src/auto_train/requirement.txt
asn1crypto==0.24.0
audioread==2.1.8
backports.functools-lru-cache==1.5
bcrypt==3.1.7
cffi==1.12.3
configobj==5.0.6
cryptography==2.7
cycler==0.10.0
cymysql==0.9.13
Cython==0.29.14
decorator==4.4.1
enum34==1.1.6
funcsigs==1.0.2
future==0.18.2
futures==3.3.0
ipaddress==1.0.22
joblib==0.14.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
numpy==1.16.5
numba==0.46.0
#matplotlib==2.2.4
pandas==0.24.2
paramiko==2.6.0
pycparser==2.19
pydub==0.23.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.1
#pyworld==0.2.8
PyYAML==5.1.1
resampy==0.2.2
scikit-learn==0.20.4
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
SoundFile==0.10.3.post1
SQLAlchemy==1.3.6
#subprocess32==3.5.4
tqdm==4.32.2
xmltodict==0.12.0
